Read status directly from FDB
